+++ packages/gatsby-theme-docs/src/components/release-notes-image.js
